Another int/long issue

+++ Velocitool/VTTrackFromDevice.h
@@ -18,7 +18,7 @@
 	VTDevice *device;
 	NSMutableArray *selectedTrackLogs;
 	NSMutableArray *trackpoints;
-	int numTrackpoints;
+	NSUInteger numTrackpoints;
 	VTCapturedTrackElement *capturedTrackXMLElement;
 		
 	NSOperationQueue *queue;
@@ -28,7 +28,7 @@
 @property (nonatomic, readonly) VTDevice *device;
 @property (nonatomic, readonly) NSMutableArray *selectedTrackLogs;
 @property (nonatomic, readwrite, retain) NSMutableArray *trackpoints;
-@property (nonatomic, readwrite) int numTrackpoints;
+@property (nonatomic, readwrite) NSUInteger numTrackpoints;
 @property (nonatomic, readwrite, retain) VTCapturedTrackElement *capturedTrackXMLElement;
 
 - (id)initWithDeviceAndTrackLogs:(VTDevice *)device trackLogs:(NSMutableArray *)trackLogs;
